Move most of ClientUtilsMixin (renamed to DataTestMixin) to a new file so that lib/tests/utils.py isn't so large.
Also add sample_image_as_file() to this class for convenience.
